new config in package, new flowchart, new name file for functions
